+++ oximod/tests/validate_required_enum.rs
@@ -1,0 +1,99 @@
+mod common;
+
+use common::init;
+use mongodb::bson::oid::ObjectId;
+use oximod::Model;
+use serde::{Deserialize, Serialize};
+use testresult::TestResult;
+
+#[derive(Model, Serialize, Deserialize, Debug)]
+#[db("test")]
+#[collection("validate_required_enum")]
+pub struct User {
+    #[serde(skip_serializing_if = "Option::is_none")]
+    _id: Option<ObjectId>,
+
+    #[validate(min_length = 5, max_length = 10)]
+    name: String,
+
+    #[validate(required, email)]
+    email: Option<String>,
+
+    #[validate(required, enum_values("admin", "user", "guest"))]
+    role: Option<String>,
+}
+
+// Run test: cargo nextest run test_missing_required_email
+#[tokio::test]
+async fn test_missing_required_email() -> TestResult {
+    init().await;
+    User::clear().await?;
+
+    let user = User {
+        _id: None,
+        name: "Valid".to_string(),
+        email: None, // ❌ required
+        role: Some("user".to_string()),
+    };
+
+    let err = user.save().await;
+    assert!(err.is_err());
+    assert!(format!("{:?}", err).contains("is required"));
+    Ok(())
+}
+
+// Run test: cargo nextest run test_missing_required_role
+#[tokio::test]
+async fn test_missing_required_role() -> TestResult {
+    init().await;
+    User::clear().await?;
+
+    let user = User {
+        _id: None,
+        name: "Valid".to_string(),
+        email: Some("user@example.com".to_string()),
+        role: None, // ❌ required
+    };
+
+    let err = user.save().await;
+    assert!(err.is_err());
+    assert!(format!("{:?}", err).contains("is required"));
+    Ok(())
+}
+
+// Run test: cargo nextest run test_invalid_enum_value
+#[tokio::test]
+async fn test_invalid_enum_value() -> TestResult {
+    init().await;
+    User::clear().await?;
+
+    let user = User {
+        _id: None,
+        name: "Valid".to_string(),
+        email: Some("user@example.com".to_string()),
+        role: Some("superadmin".to_string()), // ❌ not in allowed list
+    };
+
+    let err = user.save().await;
+    assert!(err.is_err());
+    assert!(format!("{:?}", err).contains("must be one of"));
+    Ok(())
+}
+
+// Run test: cargo nextest run test_valid_required_enum
+#[tokio::test]
+async fn test_valid_required_enum() -> TestResult {
+    init().await;
+    User::clear().await?;
+
+    let user = User {
+        _id: None,
+        name: "Valid".to_string(),
+        email: Some("user@example.com".to_string()),
+        role: Some("admin".to_string()), // ✅ allowed enum
+    };
+
+    let result = user.save().await?;
+    assert_ne!(result, ObjectId::default());
+    Ok(())
+}
